(b20859) [WIDGET] Implement Button
Implements the Button widget and modifies AppTray to use it.

Modified focus interception so that it works like touch
interception: returning `true` means that the event has been
completely consumed and that no further processing will take place --
including normal focus management.  Returning `false` means to handle
the event as normal.  In this way, AppTray now returns `true` for
everything that it doesn't want to get focus and `false` for the few
things that it does.  In addition to being more consistent, it allows
Widget.onFocus() to return to `protected`.

One challenge here was how to handle interception for a given Widget
and any *unexposed* children it might have.  This is important
because states are implemented as unexposed children, so unless
intercepting the child is the same as intercepting the parent,
interception would be broken because outside code would not be able
to discern the relationship.

As a stop-gap solution, I added the "handlesEventFor()" facility.
I'm not very happy with it, but a better solution would require more
thought and more extensive changes -- and Srini is in the middle of
making big changes in our input code, so it's better to do this after
he's done anyway. It's also not particularly pressing at the moment.
I'll file a "refactoring" ticket for it and get to it later.

Patch 2: Per Sveta's suggestion, refactored Button's internal "Plus"
classes up into TextWidget -- TextInfo is now TextParams (a la
LayoutParams), and TextWidget now caches the background and text
colors itself.  This gets rid of the extra "Plus" interface
indirection and makes the params mechanism available for general use.

Patch 3: Cleaned up whitespace, added comments to Button.setText() to
clarify the swapping done between TextWidget and TextParams.

Patch 4: D'oh! More whitespace cleanup.

Change-Id: I0c124caba9b5fb922eb59184b7a867b36bd08e87