(b21218) [WIDGET] Hoist Button.getDefaultLayout() up into Widget
Hoisted getDefaultLayout() from Button into Widget.  The best way to
invoke getDefaultLayout() was tricky to determine, but it was decided
that if mLayouts is empty when create() is called, the default layout
would be applied at that time.  The thinking here is that users
calling applyLayout() will generally
- Want to use the new layout _instead_ of the default
- Call applyLayout() at construction time, or before attaching to a
parent

If users wish to use a new Layout in _addition_ to the default, they
must either explicitly call applyLayout(getDefaultLayout())
themselves or wait until after create() is called, with onAttached()
being the obvious hook.

Some additional items:
- There is now a version of applyLayout() that takes a
`preventLayout` parameter, similar to versions of addChild()
- Since mIsCreated is only checked in the context of create(), it's
OK (and probably desirable) to check it and set it in create()
itself, rather than in the Runnable
- Renamed Button.sDefaultLayout to the non-static mDefaultLayout
- Added attribute handling to Button's NodeEntry constructor
- Override getDefaultLayout() for CheckableButton instead of
onLayout()
- Override getDefaultLayout() for CheckableGroup

Change-Id: I93fa89deb00666d50cfd4c83b2a7f7f477f9b01e