(b21237) [WIDGET] Race condition between loading state resources and changing state
There is a race condition, but it turns out to have nothing to do
with loading the state resources.  The issue is that a Widget's
parent is set *after* the call to create(), and updateState() is
called for the first time in doOnCreate() (via setLevel()).  This
happens on the GL thread, so it's possible for the update to happen
before the Widget's parent is set -- when it does, the Widget will
not correctly detect when it's supposed to be following its parent's
state, which is what's happening in this particular case.

It's been this way from the very beginning; I'm not sure why I
thought it was a good idea to call create() before setting the
parent.

Fixed by setting the parent before calling create().

Change-Id: If825b07a85bf3728dad81a478fb0152c959a3baf