(b21279) [WONDERWALL] Startup is slow, sometimes prisms
Several changes to lighten the load on the GL thread at start-up.

First, building the wonderwall content scene was moved to the main
thread.  This makes the wall sort of "pop" into existence after the
background has been shown, which can be a bit abrupt.  Also, the
double load of the list that seems to be happening is more apparent.
Both of these can be fixed seperately.

Second, the internals of ListWidget.onChanged() and ListWidget's
dataset observer's onChanged and onInvalidated are not run on the GL
thread any longer.

Third, the handling for PanelDataLocalReceiver.onReceive() was moved
to main thread.

Finally, it turns out that Class.getCanonicalName() is actually kind
of an expensive call, and we use it for loading default resources.
Added code to cache the canonical names; it doesn't save a lot, but
it was low-hanging fruit.

A few other minor tweaks:
- Removed unnecessary "public" from several interface methods
- Renamed OnAttachedListener.onAttached() to
OnAttachedToScene.onAttachedToScene() to avoid confusion with
Widget.onAttached().

Patch 2: Backed out the ListWidget GL thread changes.  Sveta was
seeing some issues, and it makes sense that there could be some
synchronization problems. It will probably require the work for
b20833 for that change to be OK.

Patch 3: Per Tom's suggestion, buildWonderWallContentScene() doesn't
need to be called on MainThread, as it's already being run on
gvrf-main thread and not on GL.  The gain was in getting ListWidget's
bits off the GL thread.

Patch 4: Cleaned up some slopiness in my code.

Patch 5: Making buildAppsWall() conditional doesn't work; we use the
list of VR apps from it in MyAppsContentScene.  No crash; but the
content scene doesn't show anything.

Change-Id: I5f865fbe6a0f07bb47c33b590c2943b0b3a5726f