(b21179) Create AppSelector Screen with all necessary components
Two significant things, plus a bunch of nit-picky stuff.

First, the new version of recycle(int dataIndex) wasn't actually
getting called.  Fixed that, and slightly tweaked it.

Second, ListWidget's overload of clear() wasn't calling super.

Of the nit-picky stuff, renaming SelectingAdapter.setEnd to setLength
and changing the parameter names is the most significant. We're not
actually directly specifying an end point, we're specifying a
distance from the start point.  Very minor, but the pedant in me
couldn't resist.

Change-Id: I6e148acbe8cbbc100fcf66379f69c74b1ab6684d