(b21426) [WONDERWALL] Audit logging
While not specifically called out in the ticket, this change
definitely falls under its purview: removing unnecessary calls to
toString().  In numerous instances this defeats logging
enable/disable by forcing formatting to occur even when the log is
disabled.  In general, toString() rarely needs to be called, in
particular when an Object is passed as a "%s" parameter to a
formatter.

One place we were taking a heavy hit on this was CacheDataSet.dump(),
which apparently gets called every frame.

Change-Id: Ia7f23ce56b8f09eaea812e3c295b1b4eaca3b26a