(b21475) NPE in Widget.setFocusEnabled
The code ListWidget.setItemFocusEnabled assumes that every host will
have a valid guest -- which should be the case.  Modified to check
for null, and to log a warning if that is not the case, rather than
NPE crashing.

ListWidget.setItemTouchable makes the same assumption, so I gave it
the same modification.

Change-Id: Icc226490b63644ba1c2a6338046dff373719d5f6