(b21180) [WONDERWALL] Handle the icon and back click in MyApps
Added an OnItemTouchListener in AppSelectorContentScene along with a
new method, onAppSelected(), in Events for forwarding that event to
clients.

Added code to MultiPageWidget to manage adding and removing
OnItemTouchListeners and OnItemFocusListeners to the pages;
otherwise, those listeners are paying attention to the pages, and not
the items in the pages.

Also:
- IconLauncher now eats its onTouch() events so they don't fall
through to the next page
- IconLauncher takes app's label as its name for debugging
- LauncherInfo exposes its label
- Revert Gravity.CENTER change on AppSelectorContentScene's
GridLayout; the resulting behavior for half-full pages is ...
unexpected.

Patch 2: Added call to super in IconLauncher.onTouch()

Change-Id: Ic12e2e2a62b2d72a7be6a8546a4c17cfadaba8b8