(b21547) [WIDGET] ListWidget unregisters all data set observers
Attempting to unregister a DataSetObserver that hasn't been
registered will result in Observable throwing an
IllegalStateException (which seems like a bit of an over-reaction).
ListWidget was working around this by just unregistering *all*
observers, but that was interfering with other observers and
kept MultiPageWidget pages from responding appropriately to DataSet
changes.

Changed ListWidget and MultiPageWidget to catch the exception
instead.

Also reenabled AppWallManager.allAppsList to be updated in a sorted
fashion.

Now you can see the items getting updated in "My Apps" when apps are
installed and uninstalled!

Change-Id: I41869287393d5587d736c324707249c1c50b98a9