(b21564) [WIDGET] PageIndicatorWidget crashes with zero indicators at construction
Modified PageIndicatorWidget to not attempt to create indicators at
construction if the count is zero. setPageCount() is modified to
detect if there were no indicators and set to the default page.

Patch 2: Per Sveta's feedback, setPageCount() will *not* set a
default page; it's the responsibility of the caller to set a new
current page.  Since default page is now only used in the
constructor, removed the field.

Change-Id: Ia828e4373916a6f91440137ca09cca099c95dea1