(b21195) [WONDERWALL] Track application launch recency
- Added new AppInfoFilters for recent Android apps and VR apps.
- Added code to AppInfoManager and LauncherInfo to track and update
the last launch time. Added code to LauncherDBHelper to upgrade
existing databases with new column.
- Extended NewApplicationObjectResult interface to take package name
on all methods; multiple files got minor updates for that.
- Modified ApplicationContentScene and OnAppLaunchedCallback to call
new AppInfoManager API to update launch time.
- Modified WonderwallSamsungPanel to use Adapters instead of Lists;
this allows the panel to respond to onChanged() notifications.
Refactored some of the code into internal class SamsungPanelGroup.
- Modified WonderwallRecentlyUsedPanel.AppManager methods to return
Adapters instead of lists.
- Added convenience method Widget.setScale(float).

Patch 2: rebased.

Patch 3: rebased.

Patch 4:

- Better structure for LauncherDBHelper.onUpgrade().
- Hacked a mechanism into Adapter for deferring dataset changed and
invalidated notifications to deal with the "screen freeze at
application launch time" issue that Sveta found.  The real cause is
that the change-tracking mechanism hasn't yet been integrated into
ListWidget and so the MultiPageStack in MyAppsContentScene is doing a
full rebuild of its items when it receives a dataset changed
notification because an application's "last launch time" has been
updated.

Patch 5: rebased.

Patch 6: Narrowed the scope of update deferral hack. There are
potential issues with deferring updates in the general case, because
the dataset itself still gets updated (or invalidated), so the
Adapter's client can't rely on the dataset to remain consistent.
There are various ways we could deal with this -- force update if any
of Adapter's getXXX() methods are called, throw exceptions -- but
they're all pretty messy, and we have other changes that need to be
made which will make this hack unnecessary.

Change-Id: I4c4e2fcf87e111a92addc6ce75dc0777ab1bdbe4