(b21621) [WIDGET] Clean up Widget's GVRSceneObject constructors
The Widget constructors that take GVRSceneObject are encouraging code
like this:

  GVRSceneObject object = new GVRSceneObject(gvrContext, w, h);
  Widget widget = new Widget(gvrContext, object);

This breaks encapsulation by allowing, for example, the transform to
be manipulated directly on the GVRSceneObject.  This ends up
bypassing Widget's mechanism for triggering requestLayout() for
transform changes.

Ultimately, both the constructors should be package access only. For
the moment, one of them must remain protected for two reasons:

1. The control bar used in AppRing2 is GVRSceneObject-based and will
take additional effort to transition entirely to Widget, and that may
not be worth the effort because the Wonderwall ControlBar may be able
to substitute.  For now, ControlBar is wrapped by a Widget.
2. WonderwallWebPanel has a need to create a GVRViewSceneObject to
manage an Android WebView.  To transition to Widget, we'll need a
Widget that wraps GVRViewSceneObject.  In the meantime,
WonderwallWebPanel's GVRViewSceneObject has been wrapped a little
more tightly in a Widget.

Also:
- Minor clean-up in NodeEntry and WidgetFactory
- Removed deprecated classes AppRing and LayoutManager. We haven't
had to refer to this code in a long while, and it's especially
irrelevant to Wonderwall, so it's time to go.  SimpleAnimationTracker
got updated because AppRing.AppTransformAnimation is now gone (the
referring method wasn't being used anyway).

Change-Id: I578480d5229fea3800862cc29f28102efaed5487