(b21568) [WONDERWALL] Load object.json from a public directory
TLDR: WonderwallSamsungPanel now can have background color specified
in metadata. The default color in the APK's default_metadata.json is
RED. Copy user_default_metadata.json to /sdcard/Documents on the
device and the "Recommendations" panel will override that with a
vomitous shade of green.  Nasty colors are for demonstration purposes
only.

Details:

Added methods to JSONHelpers for loading files from public
directories (like `Documents`), and modified Widget to load
`default_metadata.json` from /sdcard/Documents and replace/merge its
contents with the APK's default_metadata.json.

Modified the process of setting up metadata slightly: there is now a
callback, onSetupMetadata(), that can be overridden to process
metadata.  Modified getObjectMetadata() so that it always returns a
valid JSONObject, even if there is no metadata for the object, and
onSetupMetadata() is always called.  Updated several classes to
reflect this change.  I'm considering doing some renaming so the
whole thing is more clear; it feels like this should be onCreate(),
but we already have that.  Requires more thought.

One of the Widget constructors had to be modified to call
setupMetadata() itself because it provides a default texture, which
will end up over-writing any texture specified in the metadata.

Created a version of getJSONColor() that returns an ARGB packed
integer; renamed the old version to getJSONColorGl() because an array
of floats with value 0.0-1.0 is the GL way.

Change-Id: I530a5aa3cb8d15c5ec9721409a35981760f91c23