(b21613) [WONDERWALL] Expose JSON metadata parameters for Wonderwall panels
Added metadata support for most of the WonderwallPanel titlebar and
action button attributes.  More to come, but these are the simplest
and it also demonstrates the approach I'm taking to manage
attributes.  Feedback is quite welcome.

Patch 2: rebase

Patch 3: Overhauled the approach I'm taking.  The AttributesHolder
idea started getting pretty unwieldy when applied to the column
parameters.  This new approach is not as nicely encapsulated, but
it's more straightforward.

Also, moving to SDK 23 disabled our ability to read external files.
We have the permission set in the manifest, but we have no way of
handling prompting the user to give us permission.  So now
user_default_metadata.json will have to be pushed to a private data
directory.  Added load-user-default-metadata.sh to make the process
simpler.

Patch 4: rebase

Change-Id: I97915a3f15360d9a009624d351d36cd03fd63ce3