(b21568) [WONDERWALL] Load object.json from a public directory
TLDR: onSetupMetadata() was a bad idea.  The correct way to do the
work is in each class's constructor.

Details:

Undoes a good bit of what I did with the last commit for this.  The
onSetupMetadata() method was a tempting thing to do, because classes
were already overriding Widget.setupMetadata().  However, calling
into overridden methods from a parent class constructor is a
dangerous thing to do -- the overriding class hasn't actually been
constructed yet.

So the constructor of each class is the correct place to process
metadata and also gives classes more flexibility in when and how
their metadata is handled.

Also modified getObjectMetadata() so that it is only built once, and
then reused.

Patch 2: rebase

Patch 3: rebase

Patch 4: rebase

Change-Id: I6640403d8d18b6f1f9a15d290d6ca14f3093783b