(b21658) [WONDERWALL] Modifying a Filter can potentially "unregister" it
Replaced the mFilters HashMap with an ArrayList of FilterData.  In
addition to fixing this bug, it also removed a lot of look-ups.

Also: added work-around code for the issue with null objects in the
array returned by GVRPicker.pickObjects().  Eventually all our focus
management will be done using the cursor library, but in the meantime
this elimiates a pesky crash bug.

Change-Id: I901a0a75e8e4ac512a9cc26e7e1267fd4e5eae90