(b21613) [WONDERWALL] Expose JSON metadata parameters for Wonderwall panels
TLDR: Accumulate metadata for class and all superclasses.

The current implementation for metadata stops gathering information
as soon as it gets to metadata for the class or any of its
superclasses.  But we want to be able to specify defaults in
superclasses and have them apply as appropriate for deriving classes.

So modified the code to walk the hierarchy regardless of whether we
have metadata at the current inheritance level. This implementation
simply merges all the metadata into a JSONObject for each Widget
instance.  There are various ways this can be optimized, but that can
be done separately.

Change-Id: I31e7a40ef27f2472033f30e82a3c5963707d805e