(b21720) [WONDERWALL] JSON specification for backgrounds
Created a JSON schema for specifying backgrounds. Modified
BackgroundManager to load a file, backgrounds.json, built according
to this schema, and create backgrounds based on that spec.  In
addition to backgrounds.json, which contains specs for the new
Wonderwall backgrounds, I create old_backgrounds.json with specs for
the original VR Top backgrounds; those backgrounds are currently not
loaded.

Restored BackgroundManager's ability to deal with more than one
background; multiple backgrounds can be tested in non-Wonderwall mode
now.

Refactored BackgroundManager to use the String key directly instead
of an index

Modified ResourceBackground to move logic out of the switch statement
in the constructor and into the individual BackgroundResourceType
enumerations. Other minor refactorings for encapsulation and
readability.

Also:
- Fixed minor issue in AppInfoManager: if a package add has already
been processed and we get another add for the same package, we would
calculate a negative index for the package and get an
ArrayIndexOutOfBoundsException.  This is definitely a corner case,
but I did manage to cause it the other day.
- Minor cleanup in LauncherViewManager
- Moved getId() from WidgetStateInfo to Helpers so it can be used by
BackgroundManager

Change-Id: Ia7bd9df4526f5ab202ff49e6909808f15bb913b3