(b21650) Fix layout issue with dynamically changing size of the items
This isn't a fix; it is a heinous hack that alleviates the issue so
that the buttons are visible and usable in the "My Apps" and "Add"
content scenes. It does so by *hard-coding* the view port values for
the MultiPageStack and the GroupWidget containing it and the page
selector widget.  As such, it does nothing for the more general
layout issues that are present, and occassionally do odd things to
the button bar and the "Recent Apps" layout.

I'm doing this expressly to make the demo better.  The truth is, I
simply don't grok the layout and ListWidget/MultiPageStack code, or
the interaction between the two, well enough to be able to fix this
properly in any reasonable time frame.  This is definitely a bad
thing on its own, and is an issue that will need to be addressed
separately.

Also: contains minor tweaks for FilterList.

Change-Id: I11a3a1eb1f0ba1602a83f7b9807fcc6abdd22fe2