(b21886) [WIDGET] Update TextWidget and LightTextWidget to construct from JSON
TLDR: TextWidget, LightTextWidget, and Button can now be constructed
directly from JSON properties.

In order to test the feature, modified Button as well, along with
WonderwallPanel.  TextParams is separable from TextWidget, so did
that.

Added several schemas and updated existing schemas to use the new
ones.

Patch 2: Rebased and fixed format specifier error that was causing a
crash.

Patch 3: Rebased

Patch 4: Rebased. Also: Widget was not creating a default material
for "empty" scene objects, which is a problem if a mesh is later
attached and the code wants to set a texture. Apparently it's also a
problem when a loaded model has an empty sibling (WonderwallX3DPanel
and X3dWidget), though I have no idea why. Fixed that.

Added better defaulting for PointF, to support only specifying one
dimension in metadata and defaulting the other.  Modification in
user_default_metadata.json to demonstrate on Recommentations panel.

Patch 5, 6: Rebased

Patch 7: Rebased and fixed hang issue.  Turns out that having Widgets
with internal GVRSceneObject without a mesh is really bad.  That
shouldn't be the case, but it hangs the app.  My guess is that it has
to do with the infinite values that we end up generating for the
layout code.  For now, if no size is specified, we'll create a 0,0
quad, and I'll file a ticket for later investigation.  Note: this
doesn't create any more quads than we had before: this change
actually introduced having GVRSceneObjects without meshes;
previously, we were giving everything "empty" a 0,0 quad anyway.

Also a couple minor JSONException handling tweaks.

Patch 8: Fixed an issue with how TextParams was loading from JSON; it
wasn't using any existing text as a default, so if the text was set
via the constructor, it would get overwritten if there was no text
specified in the JSON properties.

Also added LightTextWidget JSONObject constructor.

Change-Id: I85247c7cf1c18dfbda93f564b0844d105061cea6