Defer initial layout until first onStep() call
This is a preliminary step toward restoring the old BoundingBox code
(which is necessary because we need more control over what is
included or not beyond "is it in the scene graph").  One issue with
restoring that code is that it iterates the list of children and at
start up, because we are adding many children, each of which requests
a layout (which iterates the children as well) a
ConcurrentModificationException is inevitable.

While this change doesn't completely eliminate that possibility, it
becomes far less likely.

Patch 2: Removed updatedFrontFacingRotationAtStart flag; no longer
needed.  Rebased.

Change-Id: I5b6cd69bbd7d9818959c3b1a18e0c57e3f665847