(b21838) 3D Asset Animation Should Only Start on Hover
Working with Mitch today, I tracked down an issue with how ListWidget
handles item focus.  By returning true, ListWidget is telling
Widget that all focus handling has been done by the listener and
bypasses and further processing of the event.  This was preventing
the event from propagating to X3DWidget, where Mitch is working to
trigger the Sensor.  So this change won't *fix* b21838, but it will
make it possible for Mitch to effect further changes that *should*
fix it.

Patch 2: rebased.

Change-Id: Ie2950297d36d2c5b0f33af1779b083ebe2962b9a