(b21879) [WIDGET] Make Widgets constructable from JSON
This reverts commit 1eea9ab82a4bd3e8f4935e2428e39885cb820572.

The issue was that if a position wasn't available in the JSON
metadata (which it would not be for any of the X3D scene object
children), the position was getting defaulted to {0, 0, 0} --
everybody to the center, please!

Also removed defaults for scaling and rotation pivot properties;
although there was no danger there, because the properties were being
checked for before attempting to get them, the defaults were
unnecessary (and in the case of scaling, wrong).

Change-Id: If1cd3ed6c30b20d0af4be24014b9685f86314971