(b21899) [WONDERWALL] [JSON] Default textures for WonderwallSamsungPanel
TLDR; default textures for WonderwallSamsungPanel are now defined in
default_metadata.json and can be overridden in
user_default_metadata.json.

This took significantly longer than expected due to numerous bugs in
JSONHelpers object merging code as well as UnmodifiableJSONObject and
UnmodifiableJSONArray.

Also factored TextureFactory out of WidgetStateInfo so that the code
can be used elsewhere.

Patch 2: rebased.

Patch 3: rebased.

Patch 4: rebased.

Patch 5: rebased.  Also "discovered" that if anything is specified
for the "groups" property of WonderwallSamsungPanel, then the coded
defaults for the second group (VR apps) won't be applied, so you end
up with a 1x3 group instead of the intended 3x3 group.  This isn't a
bug, but simply something to keep in mind about how defaults get
applied in WonderwallSamsungPanel.  At any rate, "fixed" this by
specifying 3x3 for the second group in default_metadata.json.

Change-Id: I83fc1c6a63bd712b8863c901d7061f15a49c9f72