(b22181) [Alpha] Samsung Recommendation Panel Should Show a Curated List of Recommended Apps
TLDR; loads a bunch of icons into the Recommendations panel, all of
which launch the store content scene when tapped.

Introduced a new abstraction, AdapterViewFactory, that does for the
view-related methods of Adapter what DataSet did for the data-related
methods.  This allows the views and data to be handled as discrete
components that are glued together by the new
DataSetViewFactoryAdapter class, and delegates more Adapter
boiler-plate code to base classes.

Also:
- Added DataSetAdapter.getItemId(), which was an omission
- Corrected the return type of DataSet.getItemId()
- Got rid of unnecessary MainScene parameter for Recently Used panel

Patch 2: Very weird ... static importing WonderwallSamsungPanel's
inner classes causes the build to fail, with errors in Widget.java
about "package TouchManager does not exist".  Of course it does, so
this is kind of a stumper.  At any rate, we can get around it by
explicitly scoping those classes.

Also rebased.

Patch 3: The mocked icons now have text labels, like regular app
icons. Since the icons are coming from assets, we can use mixed case
and spaces, so renamed the icons so that their names can be used as
the labels.

Fixed a bug in IconBuilder.getTileDestRect() in which integer
division could result in a ratio of zero; definitely not what was
intended, and resulted in the "Paintball" icon not appearing because
it's calculated width was zero.

Patch 4: Fixed a couple of very minor items discovered reviewing with
Sveta.

Change-Id: Idb4bd65ee40896067d198a728c9f91ab0cdd924b