(b22413) [REFACTOR] Implement Main->GL command buffer
In addition to CommandBuffer itself, this also implements a set of
abstractions for managing pools of commands -- Command, CommandPool,
and CommandFactory -- in order to keep memory thrashing and overhead
to a minimum.

Also: Added javadoc for a couple of parameters to WidgetLib.init()
and inlined mGvrContext; it's only used in init().

Patch 2: Added javadoc for new classes

Change-Id: I60355c5e7435dd6fc5a03bdcd23a8648f55780f9