(b22416) Move Widget methods to use Command Buffer, part 1
Updated RenderDataCache to use CommandBuffer.  In the process of
this, the Command/CommandFactory/CommandPool mechanism was
significantly refactored.  It is now a very generic, more contained,
and simpler system, and the command implementations themselves are
much smaller and have less boiler-plate code.  Smaller enough, in
fact, that the code from UpdateVisibilityCommand was moved back into
Widget.

Change-Id: I32816f422c208cd743dc3a7cf67072b16b6f2266