(b22502) [WIDGET] Controlbar button textures not rendered correctly on top of GVRF master
Updated TextureFactory.loadTexturesFromJSON() to load the
"main_texture" texture from the JSON spec into the "diffuseTexture"
key as well as the "main_texture" key.  We've been doing this for
some time with Widget.setTexture(), but this bit of code never got
updated.  It wasn't a problem before, but with GVRF master our "main"
textures no longer work for states without this.

Change-Id: I64a44230338795d07dd1e09d16212de1e741eeec