Update Framework lib to daily 863
In order to run on TVR.  Relies on sibling change in gvrf_launcher.

Updating the library results in GVRContext.getActivity() calls
throwing a NoSuchMethodError at runtime:

System.err  W  Caused by: java.lang.NoSuchMethodError: No virtual
method getActivity()Lorg/gearvrf/GVRActivity; in class
Lorg/gearvrf/GVRContext; or its super classes (declaration of
'org.gearvrf.GVRContext' appears in
/data/app/com.samsung.smcl.vr.gvrf_launcher-1/base.apk:classes3.dex)

Looking at the GVRf code, the method is clearly there, so I have no
idea why we're seeing this.  I resolved it by using
GVRContext.getContext() and casting to an Activity where necessary.
While it's currently valid to assume that the Context is an Activity,
this is kind of a gross thing to do.  I'll file a ticket for the
issue so that we can revert once it's resolved.

Patch 2: Rebased

Patch 3: Rebased

Change-Id: I43d46c757626fd7fdf37c78be89b2017d49d773e