(b22547) [WIDGET] Fix focus and touch issues with loaded meshes
TLDR: Added code to track all the sub-objects of loaded meshes and
add them as touchable and focusable elements as appropriate, just
like we do with mSceneObject itself.

We do this in a simple way, which precludes loaded meshes from having
sub-Widgets in them.  I'd like to update Widget and its schema later
so that we can do that, but for now we don't need that (much more
complicated to implement) feature.  Constructing Widgets from models
using WidgetFactory.createWidget(), however, will continue to create
sub-Widgets. Nobody is using this facility at the moment, and it is
considered deprecated.

Also: in the course of debugging this, I wrote a getFullName() method
that helpfully takes a GVRSceneObject and concatenates its name and
the names of its ancestors using the '.' character.  It looks like a
package name:

   GreatGrandparent.Grandparent.Parent.TheSceneObject

and is kind of a "path" to the object in the scene graph.  Now that
my debugging is done, nothing is using it, but it could be quite
helpful in the future.

Patch 2: Made handling of loaded mesh sub-objects switchable between
the old way (always create a Widget for each object) and the new
(manage sub-objects as part of the Widget for the root object),
defaulting to the old way.  Updated Widget's schema with the new
switch ("create_children").

Also:
- Changed NodeEntry's default root node class to GroupWidget
- Added Log.SUBSYSTEM.INPUT to help hush VRTop Cursor's racket
- Added and rearranged logging for TouchManager
- Cleaned up warnings in NodeEntry and WidgetFactory

Patch 3: Cleaned up white space.

Patch 4: Rebased.

Change-Id: I74b5016f1aa0629882925ad8171d4eb6a64a0788