(b22554) [WIDGET] Expose TextureFutureHelper through WidgetLib instance
Refactor TextureFutureHelper to be used as an instance through the
WidgetLib object rather than an interface to static methods.  This
removes the need to hold the cached textures in a static member.

Patch 2: GVRContext now a parameter to TextureFutureHelper CTOR so
the various methods don't need to have it as a parameter.

Change-Id: I9c6f2edd0407b5a386e1827cf981e04c1b349330